uniq specs index data.
Includes a hack to create our own instances of Gem::Version everywhere. This works around the @hash instance variable
possibly being set (and dumped) in the existing marshal data or in the YAML metadata in the incoming gem. This was
observed in a case where `gem build` was run under 1.8.7 but gemgate was running using 1.9.2.